Move code to lib folder
