+++ lib/main.js
@@ -4,7 +4,7 @@ const { join, resolve } = require('path')
 const { mkdtempSync } = require('fs')
 const { spawn } = require('child_process')
 const { app } = require('electron')
-const { version } = require('./package.json')
+const { version } = require('../package.json')
 const yargs = require('yargs/yargs')
 const ansi = require('ansi-colors')
 const symbols = require('log-symbols')
@@ -44,6 +44,7 @@ exports.main = (argv = process.argv.slice(2)) => {
 // Setup User-Data in tmp
 const userData = mkdtempSync(join(app.getPath('temp'), 'electron-mocha-'))
 app.setPath('userData', userData)
+console.log(userData)
 
 // Clean-up in a separate process so that we're not
 // removing Electron's data out from under it
@@ -60,4 +61,6 @@ app.on('quit', () => {
 // Do not quit if tests open and close windows
 app.on('window-all-closed', () => {})
 
-exports.main()
+if (require.main === module) {
+  exports.main()
+}
