+++ lib/mocha.js
