+++ lib/run.js
@@ -152,7 +152,7 @@ exports.handler = argv => {
 
         // Undocumented call in electron-window
         win._loadURLWithArgs(
-          join(__dirname, 'renderer/index.html'),
+          join(__dirname, '../renderer/index.html'),
           { ...argv, files },
           () => {
             if (argv.inspect) {
