+++ package.json
@@ -2,7 +2,7 @@
   "name": "electron-mocha",
   "version": "7.0.0",
   "description": "Mocha tests in Electron.",
-  "main": "main.js",
+  "main": "lib/main.js",
   "scripts": {
     "pretest": "standard",
     "test": "npm run test:main && npm run test:renderer",
